telegram-babushka-channel/requirements-bot.txt
aiogram>=3.4,<4
SQLAlchemy[asyncio]>=2.0
aiosqlite>=0.19
python-dotenv>=1.0
openai>=1.40
